Refactor: Move TypedDict semantic analysis to a new module
